requirements.txt
# Core dependencies
gradio>=4.0.0
streamlit>=1.28.0

# Document processing
python-docx>=1.1.0
docx2txt>=0.8

# AI/ML and RAG
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
openai>=1.0.0
anthropic>=0.8.0
google-generativeai>=0.3.0

# Vector database and embeddings
chromadb>=0.4.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Web scraping and data collection
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# File handling
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
typing-extensions>=4.8.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
